refactor(editor): complete selection and component placement hooks
